+++ packages/configs/default/index.json
@@ -26,7 +26,7 @@
     "*.pug": ["@parcel/transformer-pug"],
     "*.coffee": ["@parcel/transformer-coffeescript"],
     "*.mdx": ["@parcel/transformer-mdx"],
-    "*.{png,jpg,jpeg,webp}": ["@parcel/transformer-image-resize", "..."],
+    "*.{png,jpg,jpeg,webp}": ["@parcel/transformer-image", "..."],
     "url:*": ["@parcel/transformer-raw"]
   },
   "namers": ["@parcel/namer-default"],
