+++ packages/configs/default/package.json
@@ -54,6 +54,6 @@
     "@parcel/transformer-typescript-types": "^2.0.0-alpha.3.1",
     "@parcel/transformer-yaml": "^2.0.0-alpha.3.1",
     "@parcel/transformer-mdx": "^2.0.0-alpha.3.1",
-    "@parcel/transformer-image-resize": "^2.0.0-alpha.3.1"
+    "@parcel/transformer-image": "^2.0.0-alpha.3.1"
   }
 }
