+++ packages/transformers/image/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@parcel/transformer-image-resize",
+  "name": "@parcel/transformer-image",
   "version": "2.0.0-alpha.3.1",
   "license": "MIT",
   "publishConfig": {
